Initialize API app and create necessary files
This commit adds a new api app in our Django project. It adds the api app to the installed apps in settings.py. The necessary files for the API app, namely views.py, models.py, apps.py and admin.py are also created. This is the initial setup for developing new features in the api app.