Update data types in weather application
This commit updates the data types of the latitude and longitude fields in the 'Coordinates' class from float to string. Also, the type of the 'wind speed' field in the 'Weather' class has been changed from int to float.

This change was necessitated due to the API update where GPS coordinates are now returned as strings rather than floats. The 'wind speed' is also reported more accurately as a float. This ensures compatibility and accuracy with the incoming data from the API.